Share terminal progress rendering
Amp-Thread-ID: https://ampcode.com/threads/T-01a04764-36bf-70aa-8b26-9bbd59499581
